Checkpoint!
Make use of (localized) pygrub to find the boot kernel and ramdisk
inside the guest. For BSD guests, we mount the FS and look for a
xen kernel or then the standard kernel. We faill back to the old
method if we cannot find a kernel.

Localize guests just like we do with slicefix.

For ubuntu kernels, we have to fix the ramdisk to local the xen
block driver (which is in the ramdisk, but not used). 

Use IFBS for BW capping in dom0. Guests can still do their own link
shaping, but we cap bandwith outside the guest.

A bunch of non functional gre tunnel stuff based on the openvz code.
This is all going to come out and be replaced with openvswitch, but
want it in the repo.