Changes for XEN container loadinfo call, to support whole disk images and more efficient space usage.
When loading a whole disk image, we need to know what partition is
going to be the boot partition, since we do not get that info from
mkvnode.

Lets be more efficient with space/time; there is no reason to unzip
the image on the client until we create the container. Instead, tell
frisbee to write the raw ndz file to its lvm. Later, we unzip from
that into the target lvm. But to do that, we need to know how many
chunks the file is in advance, and that is easy to figure out in tmcd.