Check that fsync is supported on output file of imagezip.
Part of our hack strategy to detect and correct transient NFS errors
is to fsync after every file write and check for errors. However, some
devices don't support fsync. So, for example, if you were trying to
imagezip to /dev/null (to benchmark or check the resulting image size)
fsync would always fail and we would keep retrying the write forever.

There is already a "-N" option to avoid such checks and retries, but
Mike can never remember that!