Bugfix in the static_widearea_config path.
We were assuming that if static_widearea_config returned 0, it had
configured the physical $iface that findcnet was passed successfully.
But, static_widearea_config also returns 0 if the widearea config info
was present BUT if the given physical $iface did not match the widearea
config info.

So, catch that case, and all is well.

Also, fix a minor syntax bug; and add some debugging.