A big set of changes to how we create XEN guest disks.
Prior to this commit, XEN guests disks were single partition, no MBR,
the bits dumped into the lvm. This makes a snapshot of a XEN node,
look completely different then a physical disk image, especially if
users want more disk space (mkextrafs) inside the guest, and then want
to take a snapshot of that, and then run it on a physical node (which
was not possible).

With these changes, guests now use the same MBR layout as our version
two MBR, which makes them interchangeable with physical disk images.
In fact, the goal is to be able to switch back and forth as needed,
based on physical resource availability.