Changes for *remote* XEN shared nodes, as on the I2 pcpg nodes.
Since the pcpg-i2 nodes are so very flaky, lets try something that
does not require them to rebooted or imaged!

The key change is that on these remote nodes, we do not bridge the
physical control interface to the VM control interfaces. There is no
point since there are no routable IPs we can use, nor is there a
192.168 network that would be useful.

However, we still want to give the VMs their 192.168 address and we
still want multiple VMs on the same host to talk to each other, and we
we still want the VMs to be able to access the outside world with NAT.
So we still create the xenbr0 bridge, and give it the router address
(192.168.0.1). Any traffic heading out will be NAT's as normal, and
you can ssh into the VM using the physical host IP and the per VM
sshd port number.