A lot of changes for use on privately purchased InstaGeni Racks.
The GPO purchased racks all look the same so we can add the new
nodes statically from the XML file that HP sends us.

Well, for the UEN rack and other privately purchased racks, we do not
have any of that info, and so we have to figure things out ourselves.
We could use the standard newnode route without modification, but that
is a realy pain, say on a 33 node rack where the nodes land in any
order and you have no idea what the corresponding ilo is.

So I whacked this script to run from a data file we generate, which
contains the two bits of info from the top of the node (ilo passwd and
the ilo hostname), and the IPs we want to assign to the node and the
corresponding ilo.

So after turning all the nodes on and they checkin as usual, we run
this script with the -r option, which looks at the dhcpd.lease file to
find out what IP is which ILO, and the it asks each ilo via XMLRPC for
its hosts daya, which includes the ethernet MACs. I use that to find
the corresponding node in the leases file. Now I reorder the new_nodes
table so that pc1 is at the bottom and pcNN is at the top.

Now go back to the web interface and add all of the nodes.

Then run this script again (no -r), which will create all of the
management interfaces in the DB, reset the ilo, and then do the
normal initilo stuff for each node (boot order, password change, ssh
keys, etc).

Presto, done.