Changes to idle handling in ProtoGeni slivers
When a new sliver is created, they are given a relatively short
lifetime. This is the value of protogeni/initial_sliver_lifetime and
defaults to six hours.

A user may renew a sliver for up to the number of days in
protogeni/max_sliver_lifetime (defaults to 90 days), except in Utah
where it is 5 days (Emulab, Utah Rack, Utah DDC Rack).

The CM daemon idle code looks for idle slivers. An idle sliver is one
in which 50% of the physical nodes have been idle for three hours.
(protogeni/idlecheck_threshold). At this point an email message is
sent to the sliver creator.

If the sitevar protogeni/idlecheck_norenew is set, then the email
threatens to mark the sliver as unrenewable if it stays idle. Then, at
2 * protogeni/idlecheck_threshold, if the sliver is still idle, the
sliver is marked as unrenewable. No matter what the user does at this
point, he will not be able to renew the sliver and it will expire out
normally.

If protogeni/idlecheck_norenew is no set, behaviour remains as it is
now; a followup message is sent every 24 hours.

There is a new backend script called "setexpiration" that allows an
aggregate admin person to override the settings on a per-slice basis
so that users who have a need for a long running sliver do not have to
continually renew and/or bypas the max_sliver_lifetime setting. For
example:

boss> wap setexpiration -e YYYY-MM-DD mysliceurn

will extend the termination date to the given date. To restore the
default behavour:

boss> wap setexpiration -E mysliceurn

Note that idle checks are still made. To turn off idle checks for
a slice:

boss> wap setexpiration -i mysliceurn

To turn then back on:

boss> wap setexpiration -I mysliceurn