Change the semantics of "reset".
"reset" now just unmounts blockstores, it does not destroy them.
Need this so that people can include a SYSVOL blockstore in a custom image
(the only place where "reset" is used). Also, destroying blockstores would
have come as an unpleasant surprise to anyone who created a custom image
and then expected their data to still be around afterward!

Also work around a bizzare bug in BSD sed that happens in the prepare script.
That script does:

     sed -e '/# next line is swap device/,+1d' /etc/fstab

which should remove the matched comment and the line after it (the swapdev
entry). But if there are EXACTLY two additional lines after the matched line,
it would remove both of them (effectively "+2")! So if there was a mount for
a blockstore device after the swap device, prepare would remove that line too.

So in the finest tradition of "if it hurts, don't do it", the blockstore code
makes sure that it adds at least two additional lines.