Attempt to un-rot the eventsys code in frisbee.
Back in the day, when the paper was written, I had an eventsys interface
to the client to provide a syncronized interface to multiple running
instances. I rejuvenated the bits, so the eventsys control works again,
however some logic in the client has changed such that a second attempt
to run within the same client hangs. More work needed.