The parser-side of persistent blockstore support.
In parse-ns, we generate a list of accessible blockstores and put that in
the .input file. The accessiblity check right now is just that the blockstore
(actually lease) pid must match that of the experiment. This needs to be
generalized.

The blockstore set-lease command verifies that the asked-for lease matches
one of those accessible blockstores. If it does, it make sure the correct
size and other info wind up in the virt_blockstores table. Less obviously,
but of critical importance, it emits a "lease" virt_blockstore_attribute
with the correct lease index. This attributes gets converted into the
desire that is added by vtopgen to the .vtop file.