First attempt to cleanup some hack jobs.
Make a createdataset to handle dataset leases and move dataset specific
code out of approvelease and into Lease.pm (which is now Lease.pm.in as
it needs to be configured). Lease.pm still needs a bunch of OO-ification
to properly make datasets a subclass of leases. But, another day...