Better handling of locking and new functions.
Have locking handle the case where lease is destroyed while sleeping.
Be sure to refresh cached state after a sleep.

New functions to return an indication if a lease is "in use" (e.g., a
dataset lease is mapped by an experiment) and to return exptids of any
experiments using a lease.

Also, big ole comment about what the lease states mean.