Two new options to allow use with XEN VMs.
The -n option says to turn off the log file. We do not want to store
thes on the XEN host, will take too much room.

The -C option adds a circular buffer to store the last 8K of output
from the console. When a new client connects, dump the contents of
the circ buffer so the user sees context. At this point no longer
save output. When the user disconnects, start saving the output again
in the buffer, to dump at the next connection.