Make sure we don't do a partial sector write when using O_DIRECT.
Will cause a write error. This would only happen for the final range of
an image created with -f. We now pad out any partial last sector if direct
IO is specified (-f on imageunzip/frisbee) and write that to the file.

If the padding causes any problems in the future, we can always get rid of
it by truncating the file to the correct size when we are done...I think.