Bugfix for FreeNAS iSCSI (authorized initiators)
This commit is the counterpart to Dave's recent freenas-config
update to stabilize iSCSI configuration management in FreeNAS.
The code here preserves the iSCSI initiator group ID tags as
blockstores are mapped in and out.  The iSCSI target implementation
really does not like it when these initiator groups are redefined
while still in use by logical units (even if they are simply
being renumbered).

The bad part about this is it forced us to change the actual FreeNAS
code.  At some point I will submit a pull request for this and see
where it goes.