Create archive directory (expinfo) entry with unix gid of the project.
This is because at different times, users in different subgroups can
create an experiment with the same name. If the directory has the
unix group of the initial experiment with that name, then any other future
experiment with that name but in a different subgroup will not be able
to write the directory.