Return the root password hash in the jailconfig call.
This allows dom0 to set the password of the guest at creation time, so
that if something goes wrong, we can get in on the console. This also
fixes an error where on a shared node, we were returning the password
hash for the physical host. Return a per-node hash instead.

Also abstract out the various places we get read from /dev/urandom.