When calling DeleteSlice(), if a monitor process is running, then the slice is busy.  This might mean that the user will not be able to delete the slice for a long time, but we are having problems with users canceling slices before they finish setting up, and the XEN client side is not handling this very well. Note that the cleanupslice script calls GeniCM::CleanupDeadSlice() directly, which *does* kill the monitor, so admin cleanup is not affected.
Regarding the xen client side, signals can be blocked for a really
long time while a container is setting up, and so trying to kill it
fails, and bad things ripple out. Fixing that is going to take some
time to get right, so just avoid the problem for now.