Add /var/emulab/vms to the list of directories we clean.
We also make sure that all dirs in /var/emulab are really dirs and
not symlinks. Having /var/emulab/vms be a symlink (left over from
a vnode setup) caused vast confusion when trying to use the new
images for vnodes.

Note: you really, really should not image the vhost in an experiment
with vnodes. Prepare in such an environment is destructive and you
won't get a clean image anyway (due to LVM state). Swap the Xen image
in on a node directly and image that.