Fix the "dark magic" used to overwrite an existing partition.
The old way worked...sorta. When we deleted the old partition and then
recreated it with gpart, it would create the new one at the end of the
disk, not likely where the old partition was previously.

In elabinelab setup the combo:

   mkextrafs -s 2 ... /usr/testbed
   mkextrafs -s 4 ... /q

resulted in slice 2 getting created at the end of the disk (where slice 4
would normally be) and then slice 4 getting created in the only remaining
available space (where slice 2 would normally be). Very confusing to my
Monday morning brain...