Support for MBR v3 images in Xen VMs.
If the Xen dom0 base image is properly (1M) aligned and the VM images
are properly aligned as well, we should avoid or at least minimize any
anomolous effects due to mismatches along the path involving the guest OS,
guest disk layout, LVM LVs, LVM PVs, and the underlying physical disks.

Also, random change to add a min size parameter to the findSpareDisks
function. I was going to use this to avoid sucking up the little 3GB
unused partition in MBR 3, but then decided against it.