Change to use only geom and GPTs for FBSD 10 and above.
As of FreeBSD 10, I am tired of fighting the old MBR tools.
So for whole disks (slice == 0) we are going to use GPT so that we
can get good (1M) alignment and potentially big-ass partitions with
a minimum of fuss.

This means that you cannot image those partitions since imagezip
does not yet (as of 05/2014) understand GPT. But we have no mechanism
for capturing an image from anything but the system disk anyway, so
we can live with this until GPT support is added.