Add a variable to control what space is included in the xen_vg VG.
Setting LVM_FULLDISKONLY will effectively tell it to not use extra
space on the system disk when constructing the VG used for vnode disk
creation. This can affect performace for everybody, especially if it
uses multiple little partitions on the system disk. However, this is
turned off by default since, on our d710 nodes, most of the extra
space is on the system disk. Setting this should be plumbed through
to the user somehow so they can choose.