Support for a private ilo management subnet, on GENIRACKS.
The goal is to put all of the ilo management interfaces (except for
the control node of course) on to a private network. We determine
this by looking at the IP in the interface entry, but since dhcpd
will not allow a mix of subnets have to create a new subnet for it.

Then when we have to interact with the ilo we will have to set up an
openvpn tunnel to the control node. Why a VPN? Cause the remote
console will not work across an http proxy.

This code can be further generalized if we think it will be useful in
a non-genirack setting.