Change sshtb to allow "-host $BOSSNODE".
Needed when boss is the tip server (i.e., for IPMI consoles).

Somehow this vaguely feels like a potential flaw. Prior to this,
a misconfigured sshtb command that specified boss would fail.
Now it will succeed and run in the one place it can cause the
most damage.

Also, fix a couple of nits in the console code.