Support for gathering and storing Infiniband interface GUIDs in the DB.
Since GUIDs are 16 bytes and our current interface MACs are only 12 bytes,
I agonized over whether to grow the mac column to 16 bytes and just treat
it as a unique identifier (which is all we use that column for anyway).
However, in the end I just added a new guid column as there were mac columns
in a variety of other tables and it wasn't clear what the relationship was
and what I might break.

So, the newnode MFS will now report back a GUID for interfaces it recognizes
as IB (FreeBSD-specific right now). The boss-side checkin code with stash
that value in new_interfaces (and later interfaces when added). For possible
backward compat, it will also generate a MAC address from that (possibly
Mellanox-specific) so that all entries in the interfaces table will have
a MAC (yes, it should really be the other way around--all interfaces should
always have a guid).

End of story. We don't do anything else with IB right now other than stash
an interface GUID.