Call libosload directly rather than invoking os_load script.
This is not so much for efficiency but because it gives us more precise
knowledge about failures. Previously, if one node in the batch sent to
os_load failed, we didn't know which one so we had to assume all failed
and go back and reload them again. Granted, this situation was one that
"should not happen", but it does happen quite a lot, at least now when
we have flaky (IPMI) power control.

Also, brought some uniformity to the messages printed out; ie., print
a freakin timestamp already!