Client-side script for creating versioned images.
This is the replacement for create-image. Rather than puzzle out how to
maintain backward-compat syntax while allowing all the new features in
create-image, I just made a new script. When we don't care about backward
compat anymore, we can just rename the script.

Anyway, this script takes a lot more parameters (see the comment at the
top) allow creation of full or delta images and uploading via NFS or
frisbee. Since delta images require a signature file, there is new code
to handle downloading (and uploading) these files. And since their size
is measured in 10s of MB, possibly too big for the default MFS, there
is code to create a temporary new MFS to hold them. Currently that MFS
is 64MB which is big enough for our current generation of images (16GB FS)
and will work on the pc600/pc850s.

Still need to redo the boss-side of image creation (create_image, not
to be confused with create-image...yeah, I know).