Robustness work on the image upload side.
Implement idle ("lack of progress") timeout on the server.

Timeout work on the client-side too. Implement idle timeouts (not for
SSL connections yet), fixup handling of over-all timeout. In particular,
if we are reading from a pipe ('-') don't start timeout til we get some
input from the pipe.

Add option to put out progress dots on the client (cuz we are all about
dots in imagezip/frisbee).