Make sure we use -i for sendmail when we pipe into it.
This prevents a line with a single "." from meaning EOF to sendmail.
How arcane!

I discovered this when I ran a create_image and I didn't get the complete
log mailed to me. This is because create_image did a frisbee download of an
image with a single chunk, which of course printed out:

  Using Multicast 235.252.1.187
  Joined the team after 0 sec. ID is 1586355915. File is 1 chunks (963200 bytes)
  .

Fortunately, "arcane" is my middle name, so it didn't take me long to find
this...